ci: passthrough some env variables down to the vm
We spin up a VM on which we end up installing and running ansible.
If we're configuring some ansible or ara environment variables on
the job (container) itself, we want the ansible running on the VM
to inherit from those settings.

In effect, it lets us record a nested ansible run on a remote ara
server, in this case demo.recordsansible.org.